+++ .github/workflows/bench.yml
@@ -64,7 +64,7 @@ jobs:
         uses: ./.github/actions/setup
 
       - name: Bench
-        run: pnpm bench
+        run: pnpm bench:ci
         env:
           ANVIL_FORK_URL: ${{ secrets.ANVIL_FORK_URL }}
           ANVIL_FORK_BLOCK_NUMBER: 17500010
